moved assets, radio, radiogroup folders
