+++ packages/components/RadioGroup/assets/controlled_radiogroup.png
