+++ packages/components/RadioGroup/assets/disabled_item.png
