+++ packages/components/RadioGroup/assets/horizontal_stacked_radiogroup.png
