+++ packages/components/RadioGroup/assets/required_radiogroup.png
